Feature: help command. Also created new commands struct which should make it easier to add new commands and ensure all commands have run and runHelp functions. Also moved repl.go to main package.
